Tutorial Avatars and Scene Transitions
Set up some basic scripts for avatar actions in the tutorial scene. The car will move when 'E' is pressed and will detect collisions with MC for stepping consequences. The curtains and alarm clock will "wake" the MC up and send them to the door. Once they hit the door the main scene will be loaded. Result scene will be loaded once 'P' is pressed and the MC makes it to the right side of the screen.