Adjusted NPC Ignore Dist and Ball Landing Position
Shortened the distance that the check looks for the MC so it doesn't ignore NPCs when the MC is clearly not in range. Will only ignore NPCs when the MC is in view and wants to play catch. The landing position of the ball was also adjusted so it looks more natural.