Ball Hit Particles, NPC BT Framework, and MC Catch Stop
Threw in a particle system for when any character gets hit by a mean ball for added comedic effect. Created a BT framework for the NPCs when we decide what kind of behavior we want for them. The MC will now stop when walking to face and play catch with the ball kid instead of continuing walking.